+++ tutorials/tutorial_performancesettings.py
@@ -29,7 +29,7 @@ import tempfile
 
 from tutorials.utils import start, finish, create_tutorial_images, select_listelement, \
     find_and_click_picture, type_and_enter
-from tutorials.utils.platform import platform_keys
+from tutorials.utils.platform_keys import platform_keys
 
 CTRL, ENTER, WIN, ALT = platform_keys()
