+++ tutorials/tutorial_remotesensing.py
@@ -25,7 +25,7 @@ import pyautogui as pag
 
 from tutorials.utils import start, finish, create_tutorial_images, select_listelement, \
     find_and_click_picture, zoom_in
-from tutorials.utils.platform import platform_keys
+from tutorials.utils.platform_keys import platform_keys
 
 
 CTRL, ENTER, WIN, ALT = platform_keys()
