+++ tutorials/tutorial_waypoints.py
@@ -29,7 +29,7 @@ import datetime
 
 from tutorials.utils import start, finish, create_tutorial_images, select_listelement, \
     find_and_click_picture, zoom_in, panning
-from tutorials.utils.platform import platform_keys
+from tutorials.utils.platform_keys import platform_keys
 
 CTRL, ENTER, WIN, ALT = platform_keys()
