+++ tutorials/tutorial_wms.py
@@ -28,7 +28,7 @@ import pyautogui as pag
 from sys import platform
 from pyscreeze import ImageNotFoundException
 from tutorials.utils import start, create_tutorial_images
-from tutorials.utils.platform import platform_keys
+from tutorials.utils.platform_keys import platform_keys
 from tutorials.utils.picture import picture
